HPF_Aging
created the priorityBump method.

This method bumps up the priority of all Tasks in a hashMap by incrementing the Integer value of each pair, adding the Task to the appropriate queue, and removing it from it's old queue, based on the Integer associated with the Task in the HashMap.

If a Task is already at Priority 1, maxPriority, then that Task is removed from the age-tracking HashMap. If it's already maxPriority, the aging algorithm doesn't care about it.